rename: `PrefectureCheckbox` → `index`
